fix(mps): cast bfloat16 tensors to float32 before numpy conversion
Apple Silicon runs the VAE/preview models in bfloat16, but NumPy has no
native bfloat16 support, so .cpu().numpy() crashed with
"Got unsupported ScalarType BFloat16" during the live preview and again
when decoding the final image.

+++ modules/core.py
@@ -261,7 +261,7 @@ def get_previewer(model):
             x_sample = x0.to(VAE_approx_model.current_type)
             x_sample = VAE_approx_model(x_sample) * 127.5 + 127.5
             x_sample = einops.rearrange(x_sample, 'b c h w -> b h w c')[0]
-            x_sample = x_sample.cpu().numpy().clip(0, 255).astype(np.uint8)
+            x_sample = x_sample.to(torch.float32).cpu().numpy().clip(0, 255).astype(np.uint8)
             return x_sample
 
     return preview_function
@@ -335,7 +335,7 @@ def ksampler(model, positive, negative, latent, seed=None, steps=30, cfg=7.0, sa
 @torch.no_grad()
 @torch.inference_mode()
 def pytorch_to_numpy(x):
-    return [np.clip(255. * y.cpu().numpy(), 0, 255).astype(np.uint8) for y in x]
+    return [np.clip(255. * y.to(torch.float32).cpu().numpy(), 0, 255).astype(np.uint8) for y in x]
 
 
 @torch.no_grad()
